requirements.txt
accelerate==1.10.1
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
anytree==2.13.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
backoff==2.2.1
beartype==0.18.5
boto3==1.39.11
botocore==1.39.11
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
elastic-transport==8.17.1
elasticsearch==8.13.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.111.1
fastapi-cli==0.0.13
filelock==3.19.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gensim==4.3.3
graspologic==3.4.4
graspologic-native==1.2.5
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
hnswlib==0.8.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
hyppo==0.5.2
idna==3.10
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
llvmlite==0.45.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
nano-vectordb==0.0.4.3
neo4j==5.28.2
networkx==3.4.2
numba==0.62.0
numpy==1.26.4
ollama==0.6.0
openai==1.109.1
oracledb==3.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
POT==0.9.6.post1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pymilvus==2.6.2
pymongo==4.15.1
PyMySQL==1.1.2
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
redis==5.0.8
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.1.0
rich-toolkit==0.15.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
